Fix opviewer code contexts
Show the correct contract for each call. Fixes the previous assumption
that a single contract can be mapped to a given depth. This is not true
b/c if I make a call, then return, then make a 2nd call, the target of
the 1st & 2nd call will not be the same, but they will be at the same
depth