Bug and performance fixes for opviewer w/ contract sources
Changes the step from 10 to 100 for trace navigation if there are
more then 1000 opcodes.

Fixes a bug where only the source text for the last "call" op found
in the bytecode was loaded.

Reduces the lookup time once the opviewer is actually loaded. Previously
we would look up & re-create the "view" for a given opcode on every move
of the opviewer. Now we cache these "views" within the contract class
and load pre-fill the cache before displaying the opviewer. This greatly
improves performance when viewing txs with large and/or many files.

We now map the "step" to a specific contract, eliminating the need to
calculate if and how many "depth" changes occured in the last move.

Fixes a bug where the contract wasn't found for a given address due to
the last 34 bytes of the runtimeBytecode being different. This is b/c
those bytes only contain metadata (from my understanding) which can be
different for a compilation, even though the contract source is
unchanged.

Fixes a bug in the regex for pulling a single contract out of a
multi-contract file.

Simplifies the logic for highlighting the "current code"